requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.2.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   jupyter-server
    #   openai
appnope==0.1.3
    # via ipykernel
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.14.0
    # via jupyterlab-server
bcrypt==4.1.2
    # via streamlit-authenticator
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.1.0
    # via nbconvert
blinker==1.7.0
    # via streamlit
cachetools==5.3.2
    # via streamlit
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
comm==0.2.1
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distro==1.9.0
    # via openai
executing==2.0.1
    # via stack-data
extra-streamlit-components==0.1.60
    # via streamlit-authenticator
fastjsonschema==2.19.1
    # via nbformat
fqdn==1.5.1
    # via jsonschema
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
importlib-metadata==6.11.0
    # via streamlit
ipykernel==6.28.0
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
    #   qtconsole
ipython==8.19.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.1
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pydeck
json5==0.9.14
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.20.0
    # via
    #   altair
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.6.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
    #   qtconsole
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
jupyter-events==0.9.0
    # via jupyter-server
jupyter-lsp==2.2.1
    # via jupyterlab
jupyter-server==2.12.2
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.1
    # via jupyter-server
jupyterlab==4.0.10
    # via notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.25.2
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.9
    # via ipywidgets
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via nbconvert
nbclient==0.9.0
    # via nbconvert
nbconvert==7.14.0
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.9.2
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.8
    # via ipykernel
notebook==7.0.6
    # via jupyter
notebook-shim==0.2.3
    # via
    #   jupyterlab
    #   notebook
numpy==1.26.3
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.6.1
    # via -r requirements.in
overrides==7.4.0
    # via jupyter-server
packaging==23.2
    # via
    #   altair
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   qtconsole
    #   qtpy
    #   streamlit
pandas==2.1.4
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   -r requirements.in
    #   streamlit
platformdirs==4.1.0
    # via jupyter-core
prometheus-client==0.19.0
    # via jupyter-server
prompt-toolkit==3.0.43
    # via
    #   ipython
    #   jupyter-console
protobuf==4.25.1
    # via streamlit
psutil==5.9.7
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyarrow==14.0.2
    # via streamlit
pycparser==2.21
    # via cffi
pydantic==2.5.3
    # via openai
pydantic-core==2.14.6
    # via pydantic
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
    #   rich
pyjwt==2.8.0
    # via streamlit-authenticator
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
    #   pandas
    #   streamlit
python-json-logger==2.0.7
    # via jupyter-events
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   jupyter-events
    #   streamlit-authenticator
pyzmq==25.1.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   qtconsole
qtconsole==5.5.1
    # via jupyter
qtpy==2.4.1
    # via qtconsole
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via
    #   jupyterlab-server
    #   streamlit
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rich==13.7.0
    # via streamlit
rpds-py==0.16.2
    # via
    #   jsonschema
    #   referencing
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
streamlit==1.29.0
    # via
    #   -r requirements.in
    #   extra-streamlit-components
    #   streamlit-authenticator
streamlit-authenticator==0.2.3
    # via -r requirements.in
tenacity==8.2.3
    # via streamlit
terminado==0.18.0
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   streamlit
    #   terminado
tqdm==4.66.1
    # via openai
traitlets==5.14.1
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
types-python-dateutil==2.8.19.20240106
    # via arrow
typing-extensions==4.9.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
tzdata==2023.4
    # via pandas
tzlocal==5.2
    # via streamlit
uri-template==1.3.0
    # via jsonschema
urllib3==2.1.0
    # via requests
validators==0.22.0
    # via streamlit
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.7.0
    # via jupyter-server
widgetsnbextension==4.0.9
    # via ipywidgets
zipp==3.17.0
    # via importlib-metadata
